requirements.txt
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
Flask==3.1.3
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
pillow==12.2.0
requests==2.33.1
urllib3==2.6.3
Werkzeug==3.1.8
